+++ src/main/java/org/xproc/pep/ParseErrorEvent.java
@@ -1,0 +1,44 @@
+/*
+ * $Id: ParseErrorEvent.java 1807 2010-02-05 22:20:02Z scott $ 
+ * Copyright (C) 2007 Scott Martin
+ *
+ * This library is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation; either version 2.1 of the License, or (at your
+ * option) any later version. The GNU Lesser General Public License is
+ * distributed with this software in the file COPYING.
+ */
+package org.xproc.pep;
+
+
+/**
+ * An event triggered when an error occurs during parsing.
+ * @author <a href="http://www.ling.osu.edu/~scott/">Scott Martin</a>
+ * @version $LastChangedRevision: 1807 $
+ * @see ParserListener#parseError(ParseErrorEvent)
+ */
+public class ParseErrorEvent extends ParseEvent {
+	private static final long serialVersionUID = 1L;
+	
+	PepException cause;
+
+	/**
+	 * An event generated by an error during parsing at a specified index.
+	 * @param cause The exception that caused this error to be generated.
+	 * @see ParseEvent#ParseEvent(EarleyParser, Parse)
+	 */
+	ParseErrorEvent(EarleyParser earleyParser, Integer index, Parse parse,
+			PepException cause) {
+		super(earleyParser, index, parse);
+		this.cause = cause;
+	}
+
+	/**
+	 * Gets the exception that caused this error event to be generated.
+	 * @return The exception that was specified when this event was created.
+	 */
+	public PepException getCause() {
+		return cause;
+	}
+	
+}
