+++ src/main/java/org/xproc/pep/ParserEvent.java
@@ -1,0 +1,44 @@
+/*
+ * $Id: ParserEvent.java 3114 2012-02-12 19:59:59Z scott $
+ * Copyright (C) 2007 Scott Martin
+ *
+ * This library is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation; either version 2.1 of the License, or (at your
+ * option) any later version. The GNU Lesser General Public License is
+ * distributed with this software in the file COPYING.
+ */
+package org.xproc.pep;
+
+import java.util.EventObject;
+
+
+/**
+ * Abstract class for {@link EarleyParser Earley parser} events.
+ * @author <a href="http://www.ling.osu.edu/~scott/">Scott Martin</a>
+ * @version $LastChangedRevision: 3114 $
+ * @see EarleyParser
+ * @see ParserListener
+ */
+public abstract class ParserEvent extends EventObject {
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+	/**
+	 * Creates an event generated by the specified parser.
+	 * @param earleyParser The parser that generated this event.
+	 */
+	protected ParserEvent(EarleyParser earleyParser) {
+		super(earleyParser);
+	}
+	
+	/**
+	 * Gets the Earley parser that generated this event.
+	 * @return The Earley parser that was specified when this event was created.
+	 */
+	public EarleyParser getEarleyParser() {
+		return (EarleyParser)getSource();
+	}
+}
