+++ src/main/java/org/xproc/pep/ParserOptionEvent.java
@@ -1,0 +1,56 @@
+/*
+ * $Id: ParserOptionEvent.java 327 2007-04-14 18:10:55Z scott $
+ * Copyright (C) 2007 Scott Martin
+ *
+ * This library is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation; either version 2.1 of the License, or (at your
+ * option) any later version. The GNU Lesser General Public License is
+ * distributed with this software in the file COPYING.
+ */
+package org.xproc.pep;
+
+
+/**
+ * An event generated by an option being set on an {@link EarleyParser Earley 
+ * parser}.
+ * @author <a href="http://www.ling.osu.edu/~scott/">Scott Martin</a>
+ * @version $LastChangedRevision: 327 $
+ */
+public class ParserOptionEvent extends ParserEvent {
+	private static final long serialVersionUID = 1L;
+
+	ParserOption option;
+	Boolean value;
+	
+	/**
+	 * Creates a new event signaling that the specified option was set on 
+	 * the specified parser with the given value.
+	 * @param earleyParser The parser on which the option was set.
+	 * @param option The option that was set.
+	 * @param value The value assigned to <code>option</code> for this parser.
+	 */
+	ParserOptionEvent(EarleyParser earleyParser, 
+			ParserOption option, Boolean value) {
+		super(earleyParser);
+		this.option = option;
+		this.value = value;
+	}
+
+	/**
+	 * Gets the option that was set.
+	 * @return The option.
+	 */
+	public ParserOption getOption() {
+		return option;
+	}
+
+	/**
+	 * Gets the value that was set for the option.
+	 * @return The value.
+	 */
+	public Boolean getValue() {
+		return value;
+	}
+
+}
